requirements.txt
streamlit==1.38.0
pillow
